Refactor examples puppeteer test and add one for selectome.
